diploma/requirements.txt
pyTelegramBotAPI==4.5.1
google-cloud-vision==2.7.2
python-dotenv==0.20.0
